♻️ G2R0X5 task: decompose evidence command
