+++ .agentplane/tasks/202605282205-G2R0X5/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 G2R0X5 task: Evidence command decomposition [202605282205-G2R0X5]
