+++ .agentplane/tasks/202605282205-G2R0X5/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T22:06:13.021Z
+
+## Task
+
+- Task: `202605282205-G2R0X5`
+- Title: Evidence command decomposition
+- Status: DOING
+- Branch: `task/202605282205-G2R0X5/evidence-command-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282205-G2R0X5/README.md`
+
+## Verification
+
+- State: ok
+- Note: Verification passed. Commands: bunx vitest run packages/agentplane/src/commands/evidence/evidence.command.test.ts --config vitest.workspace.ts (1 file, 3 tests passed); bun run typecheck (passed); bun run arch:deps (no dependency violations); bun run lint:core (passed); bun run format:changed (passed); node scripts/checks/hotspot-report.mjs --check --warning-lines 400 --oversized-lines 600 --test-warning-lines 1000 --oversized-test-lines 1300 (passed, runtime warnings 41 -> 40).
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T22:06:13.021Z
+- Branch: task/202605282205-G2R0X5/evidence-command-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
